+++ pkg/apis/infrastructure/v1alpha2/aws_machine_deployment_types.go
@@ -6,7 +6,6 @@ import (
 
 // +genclient
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +kubebuilder:storageversion
 // +kubebuilder:subresource:status
 // +kubebuilder:resource:categories=aws;giantswarm;cluster-api
 // +k8s:openapi-gen=true