+++ pkg/apis/infrastructure/v1alpha2/g8s_control_plane_types.go
@@ -9,7 +9,7 @@ import (
 
 const (
 	kindG8sControlPlane              = "G8sControlPlane"
-	g8sControlPlaneDocumentationLink = "https://docs.giantswarm.io/reference/cp-k8s-api/g8scontrolplanes.infrastructure.giantswarm.io/"
+	g8sControlPlaneDocumentationLink = "https://docs.giantswarm.io/ui-api/management-api/crd/g8scontrolplanes.infrastructure.giantswarm.io/"
 )
 
 func NewG8sControlPlaneTypeMeta() metav1.TypeMeta {
@@ -33,7 +33,6 @@ func NewG8sControlPlaneCR() *G8sControlPlane {
 
 // +genclient
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +kubebuilder:storageversion
 // +kubebuilder:subresource:status
 // +kubebuilder:resource:categories=aws;cluster-api;giantswarm
 // +k8s:openapi-gen=true
@@ -57,7 +56,7 @@ type G8sControlPlaneSpec struct {
 	// Number of master nodes.
 	Replicas int `json:"replicas,omitempty"`
 	// Reference to a provider-specific resource. On AWS, this would be of kind
-	// [AWSControlPlane](https://docs.giantswarm.io/reference/cp-k8s-api/awscontrolplanes.infrastructure.giantswarm.io/).
+	// [AWSControlPlane](https://docs.giantswarm.io/ui-api/management-api/crd/awscontrolplanes.infrastructure.giantswarm.io/).
 	InfrastructureRef corev1.ObjectReference `json:"infrastructureRef"`
 }